+++ apps/api/package.json
@@ -33,6 +33,7 @@
     "typescript": "^5.3.3"
   },
   "devDependencies": {
+    "@types/bcrypt": "^5.0.2",
     "@types/cookie-parser": "^1.4.7",
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
